Make uid unsigned long, support full 64bit uid range (#222)

+++ src/test/scala/uk/co/gresearch/spark/dgraph/connector/partitioner/TestUidCardinalityEstimator.scala
@@ -30,7 +30,7 @@ class TestUidCardinalityEstimator extends AnyFunSpec {
       val partition = Partition(Seq.empty, Set(range))
       val actual = estimator.uidCardinality(partition)
       assert(actual.isDefined === true)
-      assert(actual.get.intValue() === range.length)
+      assert(actual.get.intValue() === range.length.intValue())
     }
 
     it("should estimate partition's uids cardinality") {
